perf: move version platform dependencies to version queries
