+++ backend/src/main/java/io/papermc/hangar/db/dao/v1/VersionsApiDAO.java
@@ -68,7 +68,14 @@ public interface VersionsApiDAO {
                                                   'external_url', pvd.external_url,
                                                   'platform', pvd.platform)) AS value
                 FROM project_version_dependencies pvd
-                WHERE pvd.version_id = pv.id)                         AS plugin_dependencies
+                WHERE pvd.version_id = pv.id)                         AS plugin_dependencies,
+               (SELECT json_agg(json_build_object(
+                   'platform', plv.platform,
+                   'version', plv.version)) AS value
+                   FROM project_version_platform_dependencies pvpd
+                    JOIN platform_versions plv ON pvpd.platform_version_id = plv.id
+                WHERE pvpd.version_id = pv.id)                        AS platform_dependencies,
+                'dum'                                                 AS platform_dependencies_formatted
            FROM project_versions pv
                JOIN project_channels pc ON pv.channel_id = pc.id
            WHERE
@@ -128,7 +135,14 @@ public interface VersionsApiDAO {
                                                   'external_url', pvd.external_url,
                                                   'platform', pvd.platform)) AS value
                 FROM project_version_dependencies pvd
-                WHERE pvd.version_id = pv.id)                         AS plugin_dependencies
+                WHERE pvd.version_id = pv.id)                         AS plugin_dependencies,
+               (SELECT json_agg(json_build_object(
+                   'platform', plv.platform,
+                   'version', plv.version)) AS value
+                   FROM project_version_platform_dependencies pvpd
+                    JOIN platform_versions plv ON pvpd.platform_version_id = plv.id
+                WHERE pvpd.version_id = pv.id)                        AS platform_dependencies,
+                'dum'                                                 AS platform_dependencies_formatted
            FROM project_versions pv
                JOIN project_channels pc ON pv.channel_id = pc.id
                <if(platformfilter)>INNER JOIN sq ON pv.id = sq.version_id<endif>
